% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WQRADMMexample.R
\name{WQRADMM}
\alias{WQRADMM}
\title{Weighted Quantile Regression for Longitudinal Big Data Based on Multi-Block ADMM
This is a non-parallel implementation.}
\usage{
WQRADMM(
  x,
  y,
  rep,
  tau,
  intercept,
  esttype,
  corrtype = "ar1",
  rhoCQR = 1,
  rhoWQR = 1,
  eps = 1e-04,
  epsw = 1e-06,
  maxstep = 3000,
  maxstepw = 100
)
}
\arguments{
\item{x}{The design matrix (without intercept)}

\item{y}{The response vector}

\item{rep}{The repeat observation number for each subject}

\item{tau}{The quantile of interest}

\item{intercept}{Whether to include the intercept into the model}

\item{esttype}{The method for estimating the regression coefficients (conventional quantile regression ("CQR") or weighted quantile regression ("WQR"))}

\item{corrtype}{The working correlation matrix to use when computing the weights, currently support ("exchangeable" and "ar1")}

\item{rhoCQR}{The augmentation parameter for ADMM when computing CQR estimator}

\item{rhoWQR}{The augmentation parameter for ADMM when computing WQR estimator}

\item{eps}{The tolerance parameter for convergence (the WQR-ADMM algorithm for computing CQR or WQR estimator)}

\item{epsw}{The tolerance parameter for convergence (the Newton-Raphson algorithm for computing the weights)}

\item{maxstep}{Maximum number of iterations allowed for the WQR-ADMM algorithm}

\item{maxstepw}{Maximum number of iterations allowed for the Newton-Raphson algorithm}
}
\value{
The coefficient estimation, the number of iterations and the running time for the WQR-ADMM algorithm, and the total time cost for computing the WQR estimator (when the parameter esttype = "WQR")
}
\description{
Weighted Quantile Regression for Longitudinal Big Data Based on Multi-Block ADMM
This is a non-parallel implementation.
}
\examples{
gcov = function(p, rho, type){
  if(type == "exchangeable"){
    cov = matrix(rho, p, p)
    diag(cov) = rep(1, p)
  }
  else{
    cov = diag(p)
    for(i in 1:p){
      for(j in 1:p){
        if(i < j) cov[i,j] = rho^{j-i}
        else cov[i,j] = cov[j,i]
      }
    }
  }
  cov
}

N = 10000
p = 100
n = 10
rep = rep(n, N)
nsum = sum(rep)
d = 0.75*p
rho_X = 0.5
rho_e = 0.5
tau = 0.75
set.seed(999)
X = matrix(rnorm(nsum*p), nsum, p)
cov_X = gcov(p, rho_X, "ar1")
X = X\%*\%chol(cov_X)
for(i in 1:d){
  X[,i] = pnorm(X[,i])
}
set.seed(1000)
e = matrix(rt(N*n, 3), N, n)
cov_e = gcov(n, rho_e, "ar1")
e = as.vector(t(e\%*\%chol(cov_e)))
sigma = 0.5
e = sigma*e
beta0 = rnorm(1)
beta = rnorm(p)
Y = beta0+X\%*\%beta+apply(X[,1:d]*e/d, 1, sum)
beta_true = c(beta0, quantile(e/d, tau)+beta[1:d], beta[(d+1):p])

WQR = WQRADMM(X, Y, rep, tau, TRUE, "WQR")
beta_WQR = WQR$Estimation_WQR
AE_WQR = sum(abs(beta_WQR-beta_true))
Time_WQR = WQR$Time_WQR
Time_WQRADMM = WQR$Time_total
}
